still need to work on router, created a context for login, will remove modal and use a transition on the same signup block instead for login
