+++ client/gui/src/App.css
@@ -19,13 +19,17 @@ body {
   height:100vh !important;
   width:100vw !important; */
   box-sizing: border-box;
-  background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' version='1.1' xmlns:xlink='http://www.w3.org/1999/xlink' xmlns:svgjs='http://svgjs.com/svgjs' width='1440' height='560' preserveAspectRatio='none' viewBox='0 0 1440 560'%3e%3cg clip-path='url(%26quot%3b%23SvgjsClipPath1035%26quot%3b)' fill='none'%3e%3crect width='1440' height='560' x='0' y='0' fill='%2332325d'%3e%3c/rect%3e%3ccircle r='22.755' cx='945.51' cy='451.3' fill='url(%23SvgjsLinearGradient1036)'%3e%3c/circle%3e%3ccircle r='37.805' cx='655.05' cy='118.66' fill='url(%23SvgjsLinearGradient1037)'%3e%3c/circle%3e%3ccircle r='37.975' cx='251.97' cy='38.68' fill='url(%23SvgjsLinearGradient1038)'%3e%3c/circle%3e%3ccircle r='16.085' cx='1367.68' cy='421.56' fill='url(%23SvgjsLinearGradient1039)'%3e%3c/circle%3e%3ccircle r='52.305' cx='769.49' cy='354.48' fill='url(%23SvgjsLinearGradient1040)'%3e%3c/circle%3e%3ccircle r='52.815' cx='1082.1' cy='208.98' fill='%2343468b'%3e%3c/circle%3e%3ccircle r='41.635' cx='1187.97' cy='353.25' fill='url(%23SvgjsLinearGradient1041)'%3e%3c/circle%3e%3ccircle r='40.87' cx='90.06' cy='77.08' fill='%2343468b'%3e%3c/circle%3e%3c/g%3e%3cdefs%3e%3cclipPath id='SvgjsClipPath1035'%3e%3crect width='1440' height='560' x='0' y='0'%3e%3c/rect%3e%3c/clipPath%3e%3clinearGradient x1='900' y1='451.3' x2='991.02' y2='451.3' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1036'%3e%3cstop stop-color='%23ab3c51' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%234f4484' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='579.4399999999999' y1='118.66' x2='730.66' y2='118.66' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1037'%3e%3cstop stop-color='%2332325d' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23424488' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='176.01999999999998' y1='38.68' x2='327.91999999999996' y2='38.68' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1038'%3e%3cstop stop-color='%23f29b7c' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%237e6286' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='1335.51' y1='421.56' x2='1399.85' y2='421.56' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1039'%3e%3cstop stop-color='%2384b6e0' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23464a8f' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='664.88' y1='354.48' x2='874.1' y2='354.48' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1040'%3e%3cstop stop-color='%2384b6e0' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23464a8f' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='1104.7' y1='353.25' x2='1271.24' y2='353.25' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1041'%3e%3cstop stop-color='%2384b6e0' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23464a8f' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3c/defs%3e%3c/svg%3e");
+  background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' version='1.1' xmlns:xlink='http://www.w3.org/1999/xlink' xmlns:svgjs='http://svgjs.com/svgjs' width='1440' height='560' preserveAspectRatio='none' viewBox='0 0 1440 560'%3e%3cg clip-path='url(%26quot%3b%23SvgjsClipPath1094%26quot%3b)' fill='none'%3e%3crect width='1440' height='560' x='0' y='0' fill='%2332325d'%3e%3c/rect%3e%3ccircle r='19.18' cx='1020.58' cy='194.14' fill='url(%23SvgjsLinearGradient1095)'%3e%3c/circle%3e%3ccircle r='26.8' cx='18.39' cy='163.29' fill='url(%23SvgjsLinearGradient1096)'%3e%3c/circle%3e%3ccircle r='16.48' cx='1202.55' cy='261.36' fill='url(%23SvgjsLinearGradient1097)'%3e%3c/circle%3e%3ccircle r='24.89' cx='1134.29' cy='199.38' fill='url(%23SvgjsLinearGradient1098)'%3e%3c/circle%3e%3ccircle r='44.055' cx='178.04' cy='117.65' fill='url(%23SvgjsLinearGradient1099)'%3e%3c/circle%3e%3ccircle r='21.395' cx='1031.22' cy='391.98' fill='url(%23SvgjsLinearGradient1100)'%3e%3c/circle%3e%3ccircle r='52.665' cx='418.91' cy='140.44' fill='%2343468b'%3e%3c/circle%3e%3ccircle r='52.395' cx='793.85' cy='319.18' fill='url(%23SvgjsLinearGradient1101)'%3e%3c/circle%3e%3c/g%3e%3cdefs%3e%3cclipPath id='SvgjsClipPath1094'%3e%3crect width='1440' height='560' x='0' y='0'%3e%3c/rect%3e%3c/clipPath%3e%3clinearGradient x1='982.22' y1='194.14' x2='1058.94' y2='194.14' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1095'%3e%3cstop stop-color='%23e298de' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23484687' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='-35.21' y1='163.29' x2='71.99000000000001' y2='163.29' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1096'%3e%3cstop stop-color='%2332325d' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23424488' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='1169.59' y1='261.36' x2='1235.51' y2='261.36' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1097'%3e%3cstop stop-color='%2384b6e0' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23464a8f' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='1084.51' y1='199.38' x2='1184.07' y2='199.38' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1098'%3e%3cstop stop-color='%23e298de' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23484687' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='89.92999999999999' y1='117.65' x2='266.15' y2='117.65' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1099'%3e%3cstop stop-color='%23e298de' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23484687' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='988.4300000000001' y1='391.98' x2='1074.01' y2='391.98' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1100'%3e%3cstop stop-color='%2384b6e0' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23464a8f' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3clinearGradient x1='689.0600000000001' y1='319.18' x2='898.6400000000001' y2='319.18' gradientUnits='userSpaceOnUse' id='SvgjsLinearGradient1101'%3e%3cstop stop-color='%23e298de' offset='0.1'%3e%3c/stop%3e%3cstop stop-color='%23484687' offset='0.9'%3e%3c/stop%3e%3c/linearGradient%3e%3c/defs%3e%3c/svg%3e");
 }
 h1{
   margin:0;
 }
+.aaa{
+  text-align: center;
+}
 h3{
-  margin:10px 0px;
+  margin:5px 0px;
+  color:rgb(198, 196, 196);
 }
 p{
   margin:0;
@@ -77,7 +81,7 @@ main{
   box-shadow: rgba(50, 50, 93, 0.25) 0px 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px;
   /*background: rgb(228, 227, 227);*/
   background:#ffffff38;
-  width:500px;
+  width:450px;
   max-height: fit-content;
   padding:30px;
   border-style: solid;
@@ -86,8 +90,6 @@ main{
 }
 img{
   /* background:rgba(128, 0, 128, 0.11); */
-  background:#800080;
-  background:url(./images/blob.svg);
   max-width: 100%;
 }
 input{
@@ -101,7 +103,7 @@ input{
 label{
   display:flex;
   flex-direction: column;
-  margin: 5px 0px;
+  margin: 15px 0px;
   font-family: 'Mulish', sans-serif;
   font-weight: 900;
 }
@@ -115,7 +117,11 @@ label{
   position: absolute;
 }
 .signup-title{
+  margin:5px 5px !important;
   padding-bottom:10px;
+  color:rgb(198, 196, 196);
+  color:White;
+  display:inline;
 }
 .requirement-warning{
   display:flex;
@@ -135,6 +141,7 @@ label{
 }
 .login-signup-link{
   font-weight: bolder;
+  cursor:pointer;
 }
 .login-forms{
   display:flex;
@@ -148,4 +155,21 @@ label{
   box-sizing: border-box;
   border-radius: 5px;
   margin:10px 0; 
+}
+input{
+  background:none;
+  border-style: none;
+  border-bottom-style: solid;
+  border-color: rgb(129, 129, 129);
+  border-bottom-width: 1px;
+  font-size:30px;
+  caret-color: white;
+}
+input:focus{
+  outline:none;
+}
+.space{
+  display:flex;
+  justify-content: center;
+  align-items: center;
 }
