+++ client/gui/src/App.js
@@ -3,11 +3,13 @@ import  { Signup } from './components/signup/Signup';
 import { Display } from './components/display';
 import { Navbar } from './components/navigation/navbar'
 import { SideImage } from './components/signup/SideImage'
+import { ModalFiles } from './components/Contexts/Modal'
 
 function App() {
 
   return (
     <div className="App">
+      <ModalFiles>
       <header>
         <Navbar/>
       </header>
@@ -19,7 +21,8 @@ function App() {
         <SideImage/>
       </div>
       </main>
-      <Display/>
+      </ModalFiles>
+      {/* <Display/> */}
     </div>
   );
 }
