+++ client/gui/src/components/signup/SideImage.jsx
@@ -1,11 +1,10 @@
 //import connect from '../../images/connect.jpg'
-import test from '../../images/test.png'
 import smol from '../../images/smol.png'
 
 export const SideImage = () =>{
     return (
         <div>
-            {<img src={smol} alt='landing-image'></img>}
+            <img src={smol} alt='landing-pic'/>
         </div>
     )
 }
